requirements.txt
autogen-agentchat==0.6.4
autogen-ext[openai]
autogen-ext[ollama]
ipykernel
python-dotenv
autogen-core
autogen-ext[http-tool]
langchain-community
autogenstudio
tiktoken
pytest
streamlit
reportlab
pdfkit
markdown
pymupdf